+++ eventsourcing-core/src/main/java/de/otto/edison/eventsourcing/KinesisEventSourceBuilder.java
@@ -5,7 +5,6 @@ import de.otto.edison.eventsourcing.consumer.EventSource;
 import de.otto.edison.eventsourcing.kinesis.KinesisEventSource;
 import de.otto.edison.eventsourcing.kinesis.KinesisStream;
 import org.slf4j.Logger;
-import org.springframework.security.crypto.encrypt.TextEncryptor;
 import software.amazon.awssdk.services.kinesis.KinesisClient;
 
 import java.util.Objects;
@@ -18,14 +17,11 @@ public class KinesisEventSourceBuilder implements EventSourceBuilder {
 
     private final ObjectMapper objectMapper;
     private final KinesisClient kinesisClient;
-    private final TextEncryptor textEncryptor;
 
     public KinesisEventSourceBuilder(final ObjectMapper objectMapper,
-                                     final KinesisClient kinesisClient,
-                                     final TextEncryptor textEncryptor) {
+                                     final KinesisClient kinesisClient) {
         this.objectMapper = objectMapper;
         this.kinesisClient = kinesisClient;
-        this.textEncryptor = textEncryptor;
     }
 
     @Override
@@ -33,7 +29,7 @@ public class KinesisEventSourceBuilder implements EventSourceBuilder {
         Objects.requireNonNull(streamName, "stream name must not be null");
         LOG.info("Building '{}' as KinesisEventSource", streamName);
         final KinesisStream kinesisStream = new KinesisStream(kinesisClient, streamName);
-        return new KinesisEventSource(name, kinesisStream, textEncryptor, objectMapper);
+        return new KinesisEventSource(name, kinesisStream, objectMapper);
     }
 
 }