+++ eventsourcing-core/src/main/java/de/otto/edison/eventsourcing/annotation/EventSourceConsumerBeanPostProcessor.java
@@ -14,11 +14,12 @@ import org.springframework.core.Ordered;
 import org.springframework.core.annotation.AnnotationUtils;
 
 import java.lang.reflect.Method;
-import java.util.*;
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.Map;
+import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 
-import static java.lang.String.format;
-import static java.util.Arrays.stream;
 import static org.slf4j.LoggerFactory.getLogger;
 import static org.springframework.core.MethodIntrospector.selectMethods;