+++ eventsourcing-core/src/main/java/de/otto/edison/eventsourcing/configuration/KinesisConfiguration.java
@@ -8,8 +8,6 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.security.crypto.encrypt.Encryptors;
-import org.springframework.security.crypto.encrypt.TextEncryptor;
 import software.amazon.awssdk.core.auth.AwsCredentialsProvider;
 import software.amazon.awssdk.core.regions.Region;
 import software.amazon.awssdk.services.kinesis.KinesisClient;
@@ -40,15 +38,8 @@ public class KinesisConfiguration {
 
     @Bean
     public KinesisEventSenderFactory kinesisEventSenderFactory(ObjectMapper objectMapper,
-                                                               TextEncryptor textEncryptor,
                                                                KinesisClient kinesisClient) {
-        return new KinesisEventSenderFactory(objectMapper, textEncryptor, kinesisClient);
-    }
-
-    @Bean
-    @ConditionalOnMissingBean(TextEncryptor.class)
-    public TextEncryptor textEncryptor() {
-        return Encryptors.noOpText();
+        return new KinesisEventSenderFactory(objectMapper, kinesisClient);
     }
 
 }